feat: log success and failures separately
This allows us to compile a list of all successes and all failures to present to the user at the end of the process. Since the logs for failures also contain the error that occurred, it is better to set these in the list of errors instead of just the numbers to which a message could not be sent.

Signed-off-by: ksdfg <ksdfg123@gmail.com>